calendar-event-dialog.xul/.js: create again these files to manage window menu and toolbar
Now that the main dialog is included in an iframe, we have to communicate
with the iframe from the parent to give to the iframe the good commands.

The structure of sendMessage/receiveMessage come from the Lightning project.
Especially the sendMessage() function which send command with the good event id
(needed for tab view).