lightning-item-iframe.js: updateScreen() shows always HTML content editor
exchangecalendar <4.0 did a test on existence of an element with id
"todo-entrydate" to decide whether it should display or hide the
HTML content editor.

(this code was in the onLoad() function of calendar-event-dialog.js)

Because of the name of the ID, I misunderstood the test and thought
the HTML editor has to be displayed only for tasks.

This commit fixes the dialog display.
The next one will fix the content saving.