mivExchangeCalendar.js: getItemsFromMemoryCache: fix looking for items in cache
The code before were using simple day index to go through start and end of the
range. Furthermore, it didn't managed that range were given in local timezone
and memory cache is in UTC.

This fix use Lightning tools to go through range and it converts local timezone
to UTC, so the cache is correctly read.

Fixes #67