Address book search contact update
When running search for contact from Exchange Calendars, we maintain now
two lists in result object:
- first list contains all contacts found saved as Cards
- second one contains, in order of reception, the localId of each Card received

So, the algorithm is efficient for search by Card (on insert) and search by Index:
we never has to loop on every elements to find what we need.

Furthermore, we have back the "o" icon in result list and we replaced
label "Exchange Contact" by the name of extension "Exchange Calendar".