Fix adding times to all day event
When an all day event is modified to have start/end times
(i.e. no longer being an all day event), the order of the transmitted
properties matters for OWA. So the all day property has to be transmitted
before the start/end times, as otherwise the desired times are not applied.

Closes #286.